requirements.txt
absl-py==0.2.2
aniso8601==3.0.0
astor==0.6.2
bleach==1.5.0
click==6.7
Flask==1.0.2
flask-restplus==0.10.1
gast==0.2.0
grpcio==1.13.0
gunicorn==19.9.0
h5py==2.9.0
html5lib==0.9999999
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==2.6.11
MarkupSafe==1.0
numpy==1.14.5
pandas==0.24.2
Pillow==6.1.0
protobuf==3.6.0
python-dateutil==2.8.0
pytz==2018.4
PyYAML==3.12
scipy==1.3.0
six==1.11.0
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
Werkzeug==0.14.1
